[Xamarin.Android.Linker] exploring the new .NET linker
Based on the new `Microsoft.NET.ILLink` NuGet package, this is a start
to what it would look like to use it for Xamarin.Android.
