+++ src/Xamarin.Android.Build.Tasks/Utilities/MonoAndroidHelper.cs
@@ -17,13 +17,12 @@ using Microsoft.Build.Utilities;
 
 namespace Xamarin.Android.Tasks
 {
-	public class MonoAndroidHelper
+	public partial class MonoAndroidHelper
 	{
 		static Lazy<string> uname = new Lazy<string> (GetOSBinDirName, System.Threading.LazyThreadSafetyMode.PublicationOnly);
 
 		// Set in ResolveSdks.Execute();
 		// Requires that ResolveSdks.Execute() run before anything else
-		public static string[] TargetFrameworkDirectories;
 		public static AndroidVersions   SupportedVersions;
 		public static AndroidSdkInfo    AndroidSdk;
 
@@ -309,32 +308,6 @@ namespace Xamarin.Android.Tasks
 			return bool.TryParse (hasReference, out bool value) && value;
 		}
 
-		public static bool IsFrameworkAssembly (string assembly)
-		{
-			return IsFrameworkAssembly (assembly, false);
-		}
-
-		public static bool IsFrameworkAssembly (string assembly, bool checkSdkPath)
-		{
-			if (IsSharedRuntimeAssembly (assembly)) {
-#if MSBUILD
-				bool treatAsUser = FrameworkAssembliesToTreatAsUserAssemblies.Contains (Path.GetFileName (assembly));
-				// Framework assemblies don't come from outside the SDK Path;
-				// user assemblies do
-				if (checkSdkPath && treatAsUser && TargetFrameworkDirectories != null) {
-					return ExistsInFrameworkPath (assembly);
-				}
-#endif
-				return true;
-			}
-			return TargetFrameworkDirectories == null || !checkSdkPath ? false : ExistsInFrameworkPath (assembly);
-		}
-
-		public static bool IsSharedRuntimeAssembly (string assembly)
-		{
-			return Array.BinarySearch (Profile.SharedRuntimeAssemblies, Path.GetFileName (assembly), StringComparer.OrdinalIgnoreCase) >= 0;
-		}
-
 		public static bool IsReferenceAssembly (string assembly)
 		{
 			using (var stream = File.OpenRead (assembly))
@@ -349,16 +322,6 @@ namespace Xamarin.Android.Tasks
 				}
 				return false;
 			}
-		}
-
-		public static bool ExistsInFrameworkPath (string assembly)
-		{
-			return TargetFrameworkDirectories
-					// TargetFrameworkDirectories will contain a "versioned" directory,
-					// e.g. $prefix/lib/xamarin.android/xbuild-frameworks/MonoAndroid/v1.0.
-					// Trim off the version.
-					.Select (p => Path.GetDirectoryName (p.TrimEnd (Path.DirectorySeparatorChar)))
-					.Any (p => assembly.StartsWith (p));
 		}
 
 		public static bool IsForceRetainedAssembly (string assembly)
@@ -548,21 +511,6 @@ namespace Xamarin.Android.Tasks
 			return ret;
 		}
 #endif
-
-		internal static readonly string [] FrameworkEmbeddedJarLookupTargets = {
-			"Mono.Android.Support.v13.dll",
-			"Mono.Android.Support.v4.dll",
-			"Xamarin.Android.NUnitLite.dll", // AndroidResources
-		};
-		internal static readonly string [] FrameworkEmbeddedNativeLibraryAssemblies = {
-			"Mono.Data.Sqlite.dll",
-			"Mono.Posix.dll",
-		};
-		internal static readonly HashSet<string> FrameworkAssembliesToTreatAsUserAssemblies = new HashSet<string> (StringComparer.OrdinalIgnoreCase) {
-			"Mono.Android.Support.v13.dll",
-			"Mono.Android.Support.v4.dll",
-			"Xamarin.Android.NUnitLite.dll",
-		};
 
 		public static Dictionary<string, string> LoadAcwMapFile (string acwPath)
 		{
